fix: Unify Integration and Unit tests in same project to run coverage in Sonar Cloud
