+++ odin-baseline-api.sln
@@ -13,8 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Odin.Baseline.Domain", "src
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Odin.Baseline.Application", "src\Odin.Baseline.Application\Odin.Baseline.Application.csproj", "{FDA70DC4-EC16-4ACD-9A4F-6006296B6FBD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Odin.Baseline.IntegrationTests", "tests\Odin.Baseline.IntegrationTests\Odin.Baseline.IntegrationTests.csproj", "{F2FBE90B-9CD6-43AD-897B-859FC78F1422}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Odin.Baseline.Infra.Data.EF", "src\Odin.Baseline.Infra.Data.EF\Odin.Baseline.Infra.Data.EF.csproj", "{077D0CCD-2CA9-49E1-ACE6-BD89872329A1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Odin.Baseline.EndToEndTests", "tests\Odin.Baseline.EndToEndTests\Odin.Baseline.EndToEndTests.csproj", "{09A7B866-7F20-47C5-80E6-09DDEF624104}"
@@ -46,10 +44,6 @@ Global
 		{FDA70DC4-EC16-4ACD-9A4F-6006296B6FBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FDA70DC4-EC16-4ACD-9A4F-6006296B6FBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FDA70DC4-EC16-4ACD-9A4F-6006296B6FBD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F2FBE90B-9CD6-43AD-897B-859FC78F1422}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F2FBE90B-9CD6-43AD-897B-859FC78F1422}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F2FBE90B-9CD6-43AD-897B-859FC78F1422}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F2FBE90B-9CD6-43AD-897B-859FC78F1422}.Release|Any CPU.Build.0 = Release|Any CPU
 		{077D0CCD-2CA9-49E1-ACE6-BD89872329A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{077D0CCD-2CA9-49E1-ACE6-BD89872329A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{077D0CCD-2CA9-49E1-ACE6-BD89872329A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -74,7 +68,6 @@ Global
 		{C2E8CCE5-5C94-49A8-90F9-5A5C4CDBDDEB} = {F703986E-B21B-4605-942D-2E43B25C91C0}
 		{4EA11FB2-7BAD-4946-8D85-F202BD675FF0} = {B93CF8B4-43EE-47CF-86DD-6AB933FF463A}
 		{FDA70DC4-EC16-4ACD-9A4F-6006296B6FBD} = {B93CF8B4-43EE-47CF-86DD-6AB933FF463A}
-		{F2FBE90B-9CD6-43AD-897B-859FC78F1422} = {F703986E-B21B-4605-942D-2E43B25C91C0}
 		{077D0CCD-2CA9-49E1-ACE6-BD89872329A1} = {B93CF8B4-43EE-47CF-86DD-6AB933FF463A}
 		{09A7B866-7F20-47C5-80E6-09DDEF624104} = {F703986E-B21B-4605-942D-2E43B25C91C0}
 		{71CB22AA-DE77-4091-B2CA-AD7061202C3C} = {B93CF8B4-43EE-47CF-86DD-6AB933FF463A}
